backend/api/requirements.txt
textract 
fastapi 
sentence-transformers
pinecone-client